[Fix] Divide files
- `DictionaryHandler.ts`
- `types.ts`
