+++ src/types.ts
@@ -1,0 +1,9 @@
+/**
+ * i18n-Lens
+ * Copyright (c) 2022-present NAVER Corp.
+ * Apache License v2.0
+ */
+
+export type Dictionary = {
+  [key: string]: string;
+};
